Bump min Perl to 5.12 to match Test::Deep
All of RJBS' dists are now 5.12+, therefore if we want to keep depending
on Test::Deep we need to be too. Personally I think 5.12 is plenty old
enough.

While here make use of one new Perl feature to declare VERSION in the
package line.

This should make the CI green again!